WIP: Partial support for non-uniform coordinate boundaries. DataSource indexes non-uniform boundaries during evaluation as needed, but interpolation raises an exception for non-uniform or non-centered boundaries.
